Processed feedback on fine-tuning Double Ratchet algorithm.
Processed the feedback. IIUC one problem still exists: we now only
increment `since_last_dh` on the instances where we rotate the brace
key. On the instances where we generate a new DH key pair, we reset from
3 to 0. However, this would effectively give us 4 distinct values: 0, 1,
2, 3. An even number would imply that DH ratchet always happens on the
same side. AFAICT this is not a desirable property. Instead, it should
be alternating, i.e. an odd number of values.